design/example/resources/simple: move exports design to separate file so we can regenerate the file
