+++ design/example/resources/simple/exports_test.go
